+++ src/templates/mini-program.ts
@@ -8,24 +8,6 @@ export default class TemplateMiniProgram extends BaseClass {
   description = 'template for mini-program application by use taro'
   templates = []
   features = []
-  copyFiles = [
-    '.gitignore',
-    '.editorconfig',
-    '.prettierignore',
-    'babel.config.js',
-    'config/*'
-  ]
-
-  renderFiles = [
-    'babel.config.js',
-    'global.d.ts',
-    'package.json',
-    'project.config.json',
-    'README.md',
-    'tsconfig.json'
-  ]
-
-  renderFileTypes = 'js,jsx,ts,tsx,css,scss,sass,less,md,vue,html'
 
   constructor(public factory: Factory) {
     super(factory)